consolidate timer functions
